Replaced float with double

+++ FIR.cpp
@@ -1,9 +1,9 @@
 #include "FIR.h"
 
-FIRfilter::FIRfilter(const float * coef, uint8_t size){
-    len = (size / sizeof(float))-1;
+FIRfilter::FIRfilter(const double * coef, uint8_t size){
+    len = (size / sizeof(double))-1;
     h = coef;
-    x = new float[len];
+    x = new double[len];
     for(i = 0; i < len; ++i){
         x[i]=0;
     }
@@ -13,7 +13,7 @@ FIRfilter::FIRfilter(const float * coef, uint8_t size){
 FIRfilter::FIRfilter(const FIRfilter &copy){
     len = copy.len;
     h = copy.h;
-    x = new float[len];
+    x = new double[len];
     for(i = 0; i < len; ++i){
         x[i]=0;
     }
@@ -25,7 +25,7 @@ FIRfilter::~FIRfilter(void){
     // Assume h is constant and doesn't need deletion
 }
 
-void FIRfilter::Reset(float val=0.){
+void FIRfilter::Reset(double val=0.){
     // Reset the internal state
     for(i = 0; i < len; ++i){
         x[i]=val;
@@ -33,8 +33,8 @@ void FIRfilter::Reset(float val=0.){
     i = 0;
 }
 
-float FIRfilter::Process(float sample){
-    float out = 0;
+double FIRfilter::Process(double sample){
+    double out = 0;
     uint8_t j = 1;
     uint8_t k;
     for(k=i; k<len; --k){
