+++ IIR.cpp
@@ -1,8 +1,8 @@
 #include "IIR.h"
 
-float IIRfilter::Process(float sample){
-    float in  = 0;
-    float out = 0;
+double IIRfilter::Process(double sample){
+    double in  = 0;
+    double out = 0;
     uint8_t j = 2;
     uint8_t k;
     uint8_t len2 = len * 2 + 2;
@@ -28,8 +28,8 @@ float IIRfilter::Process(float sample){
     return out;
 }
 
-float IIRfilter::Stability(void){
-    float a = 0;
+double IIRfilter::Stability(void){
+    double a = 0;
     // Find the total gain of the feedback.
     for(i = 2; i < (2*len); i+=2){
         a -= h[i];
@@ -38,8 +38,8 @@ float IIRfilter::Stability(void){
 }
 
 /** FIXME: I don't think this works properly */
-void IIRfilter::Reset(float val=0.){
-    float a = Stability();
+void IIRfilter::Reset(double val=0.){
+    double a = Stability();
     // Calculate the steady-state condition of a one-element feedback
     // steady = input + steady*(a) ===> steady = input / (1-a)
     val = val / (1-a);
@@ -50,10 +50,10 @@ void IIRfilter::Reset(float val=0.){
     i = 0;
 }
 
-IIRfilter::IIRfilter(const float * coef, uint8_t size){
-    len = size / (2*sizeof(float)) - 1;
+IIRfilter::IIRfilter(const double * coef, uint8_t size){
+    len = size / (2*sizeof(double)) - 1;
     h = coef;
-    x = new float[len];
+    x = new double[len];
     for(i = 0; i < len; ++i){
         x[i]=0;
     }
@@ -63,7 +63,7 @@ IIRfilter::IIRfilter(const float * coef, uint8_t size){
 IIRfilter::IIRfilter(const IIRfilter &copy){
     len = copy.len;
     h = copy.h;
-    x = new float[len];
+    x = new double[len];
     for(i = 0; i < len; ++i){
         x[i]=0;
     }
